+++ public/locales/en/translation.json
@@ -12,7 +12,7 @@
     "p1": {
       "l1": "I’m    Saumya  ",
       "l2": "A builder from heart and a systems thinker",
-      "l3": ""
+      "l3": "Plan like the Military"
     },
     "p2": {
       "l1": "Plan like the Military",
@@ -71,7 +71,7 @@
       "Now, I apply this curiosity to tech, business, mental models, history, and geopolitics.<br/> it changes how people think or operate, count me in.",
       "Give me a lever long enough<br />and a place to stand<br />and I will move the Earth - Aristotle"
     ],
-    "method2_title": "Give me a lever long enough<br />and a place to stand<br />and I will move the Earth - Aristotle",
+    "method2_title": "<br/><br/><br/>My Recommendations",
     "method2": [
       "",
       "",
