+++ src/partials/About/index.module.css
@@ -17,6 +17,14 @@
   }
 }
 
+.video {
+  width: 100%;
+  max-width: 1080px;
+  aspect-ratio: 1080 / 560;
+  object-fit: cover;
+  display: block;
+  margin: 0 auto; /* centers the video */
+}
 .awardsTitle {
   margin-top: var(--spc-10);
 }
