--- src/partials/About/slider.module.css
@@ -1,114 +1,0 @@
-.sliderContainer {
-    position: relative;
-    width: 100%;
-    height: 60vh;
-    min-height: 400px;
-    margin: var(--content-block-margin-top) 0;
-    overflow: hidden;
-    background-color: rgba(255, 255, 255, 0);
-    border-radius: 8px;
-    cursor: pointer;
-  }
-  
-  .threeContainer {
-    width: 100%;
-    height: 100%;
-  }
-  
-  .controls {
-    position: absolute;
-    bottom: 40px;
-    left: 0;
-    right: 0;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 var(--spc-4);
-    z-index: 2;
-  }
-  
-  .controlButton {
-    background: rgba(255, 255, 255, 0.15);
-    border: none;
-    width: 50px;
-    height: 50px;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    color: var(--color-white);
-    font-size: 24px;
-    transition: background-color 0.3s ease;
-    opacity: 0.7;
-  }
-  
-  .controlButton:hover {
-    background: rgba(255, 255, 255, 0.3);
-    opacity: 1;
-  }
-  
-  .slideIndicators {
-    display: flex;
-    gap: var(--spc-2);
-  }
-  
-  .indicator {
-    width: 12px;
-    height: 12px;
-    border-radius: 50%;
-    background-color: rgba(255, 255, 255, 0.3);
-    cursor: pointer;
-    transition: all 0.3s ease;
-  }
-  
-  .indicator.active {
-    background-color: var(--color-white);
-    transform: scale(1.2);
-  }
-  
-  .caption {
-    position: absolute;
-    bottom: 15px;
-    left: 0;
-    right: 0;
-    text-align: center;
-    color: var(--color-white);
-    font-size: 18px;
-    padding: 0 var(--spc-4);
-    font-weight: 500;
-    text-shadow: 0 1px 3px rgba(255, 255, 255, 0);
-  }
-  
-  .pauseIndicator {
-    position: absolute;
-    top: 20px;
-    right: 20px;
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    background-color: rgba(249, 249, 249, 0);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: var(--color-white);
-    font-size: 16px;
-    animation: fadeIn 0.3s ease;
-  }
-  
-  @keyframes fadeIn {
-    from { opacity: 0; }
-    to { opacity: 1; }
-  }
-  
-  @media (--md-l) {
-    .sliderContainer {
-      height: 70vh;
-    }
-    
-    .controlButton {
-      width: 60px;
-      height: 60px;
-      font-size: 28px;
-    }
-  }
