--- src/partials/Contact2/index.module.css
@@ -1,88 +1,0 @@
-.root {
-  @mixin body-2;
-  font-size: 20px;
-  position: relative;
-  z-index: 1;
-  height: 50vh;
-  padding-top: 50vh;
-
-  &,
-  a {
-    color: var(--color-white);
-  }
-
-  a {
-    position: relative;
-    cursor: none;
-    text-decoration: none;
-  }
-}
-
-.section {
-  margin: var(--spc-2) 0;
-}
-
-.emailContainer {
-  overflow: hidden;
-}
-
-.email {
-  display: inline-block;
-
-  &::before {
-    position: absolute;
-    z-index: -1;
-    top: auto;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    height: 0;
-    background: rgb(11 19 172);
-    background: linear-gradient(0deg, rgb(11 19 172 / 100%) 0%, rgb(151 1 152 / 100%) 100%);
-    content: '';
-    transform: translate3d(0, 100%, 0) scale(1, 1);
-    transform-origin: center bottom;
-    transition: all 0.3s ease;
-    will-change: transform;
-  }
-
-  .isEmailVisible & {
-    &::before {
-      height: 50%;
-      transform: translate3d(0, 0, 0) scale(1, 1);
-      transition: transform 0.7s ease, height 0.7s ease 1.5s;
-    }
-
-    &:hover {
-      &::before {
-        transform:translate3d(0, 0, 0) scale(1, 1.5);
-        transition: all 0.3s ease;
-      }
-    }
-  }
-}
-
-
-.list {
-  padding-top: var(--spc-5);
-  padding-left: 0;
-
-  @media (--md-l) {
-    padding-top: 0;
-    padding-left: var(--spc-10);
-  }
-
-  li {
-    list-style: none;
-
-    a {
-      @mixin body-1;
-
-      line-height: 2.5;
-      @media (--md-l) {
-        line-height: 1.7;
-      }
-    }
-  }
-}
